Add files via upload
I added the mana tower card. I also altered the executioner card so that he wouldn't add execute as a possible move if it was an invalid square, since you made it so failing to execute or executing an empty square wouldn't waste mana or exhaust the executioner, making it redundant to try. I also added the "show" option, which prints out a description of the current board.